requirements.txt
requests>=2.31.0
python-whois>=0.9.4
beautifulsoup4>=4.12.2
tldextract>=3.6.0
